Cleanup, add bean validation, add error handling

+++ image-service/src/main/java/com/paoperez/imageservice/ImageService.java
@@ -1,34 +1,38 @@
 package com.paoperez.imageservice;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
-public class ImageService {
+class ImageService {
     @Autowired
     private ImageRepository imageRepository;
 
-    public List<Image> getAllImages() {
+    List<Image> getAllImages() {
         return imageRepository.findAll();
     }
 
-    public Optional<Image> getImage(String id) {
-        return imageRepository.findById(id);
+    Image getImage(String id) {
+        return imageRepository.findById(id).orElseThrow(() -> new ImageNotFoundException(id));
     }
 
-    public Image createImage(Image image) {
+    Image createImage(Image image) {
         return imageRepository.save(image);
     }
 
-    public Image updateImage(Image image) {
+    Image updateImage(Image image) {
+        if (!imageRepository.existsById(image.getId()))
+            throw new ImageNotFoundException(image.getId());
         return imageRepository.save(image);
     }
 
-    public void deleteImage(String id) {
+    Boolean deleteImage(String id) {
+        if (!imageRepository.existsById(id))
+            throw new ImageNotFoundException(id);
         imageRepository.deleteById(id);
+        return true;
     }
 
 }
